fix(oauth): publish RFC 8414 authorization server metadata
Discovery doc was at /.well-known/openid-configuration but we don't have
an authorization_endpoint (Device Flow only). RFC 8414 is the canonical
location for OAuth-only providers; carry device_authorization_endpoint
explicitly.

Shared builder in lib/oauth/metadata.ts; both routes return identical
bodies so they can't drift. OIDC path stays for back-compat with the
already-shipped CLI and openid-client default discovery.
