+++ frontend/app/.well-known/openid-configuration/route.ts
@@ -1,24 +1,8 @@
-import { getIssuer } from "@/lib/oauth/jwt";
+import { buildAuthorizationServerMetadata, metadataResponse } from "@/lib/oauth/metadata";
 
+// Back-compat alias for already-shipped CLIs and openid-client default
+// discovery. Canonical location for OAuth-only servers is RFC 8414 at
+// `/.well-known/oauth-authorization-server`.
 export async function GET(): Promise<Response> {
-  const issuer = getIssuer();
-  const body = {
-    issuer,
-    device_authorization_endpoint: `${issuer}/oauth/device/authorize`,
-    token_endpoint: `${issuer}/oauth/token`,
-    jwks_uri: `${issuer}/oauth/jwks`,
-    grant_types_supported: ["urn:ietf:params:oauth:grant-type:device_code", "refresh_token"],
-    response_types_supported: [],
-    id_token_signing_alg_values_supported: ["RS256"],
-    token_endpoint_auth_methods_supported: ["none"],
-    scopes_supported: ["projects:rw"],
-  };
-  return new Response(JSON.stringify(body), {
-    status: 200,
-    headers: {
-      "content-type": "application/json",
-      "cache-control": "public, max-age=600",
-      "access-control-allow-origin": "*",
-    },
-  });
+  return metadataResponse(buildAuthorizationServerMetadata());
 }
